examples/integrations/gateway-mode/python/requirements.txt
# AxonFlow SDK
axonflow>=0.2.0

# LLM Providers
openai>=1.6.0
anthropic>=0.8.0

# Environment management
python-dotenv>=1.0.0
